+++ packages/create-pantheon-decoupled-kit/__tests__/handlebarsHelpers.test.ts
@@ -12,3 +12,20 @@ describe('pkgNameHelper', () => {
 		expect(result).toEqual('invalid-package-name-test');
 	});
 });
+
+describe('wpGraphqlHelper', () => {
+	it('should append /wp/graphql on to the supplied string', () => {
+		const input = 'https://my-cms-endpoint.pantheonsite.io';
+		const result = hbsHelpers.wpGraphql(input);
+		expect(result).toEqual(
+			'https://my-cms-endpoint.pantheonsite.io/wp/graphql',
+		);
+	});
+	it('should not append if the string already ends with /wp/graphql', () => {
+		const input = 'https://my-cms-endpoint.pantheonsite.io/wp/graphql/';
+		const result = hbsHelpers.wpGraphql(input);
+		expect(result).toEqual(
+			'https://my-cms-endpoint.pantheonsite.io/wp/graphql/',
+		);
+	});
+});